# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SPPD API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/jabatans": {
            "get": {
                "tags": [
                    "Jabatan"
                ],
                "summary": "Get all jabatan",
                "operationId": "a1ae0c35f65c683b9f2d3842eea78e2f",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Jabatan"
                ],
                "summary": "Create jabatan",
                "operationId": "f630d950ade510ee0c3240e9369d1128",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "department_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Manager IT"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "MGR_IT"
                                    },
                                    "level": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jabatans/{id}": {
            "get": {
                "tags": [
                    "Jabatan"
                ],
                "summary": "Get detail jabatan",
                "operationId": "69238943460d801e88a7cb4b9accdcc6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Jabatan"
                ],
                "summary": "Update jabatan",
                "operationId": "ddfaffd7d4fd2082591a5852339e6fca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "level": {
                                        "type": "integer"
                                    },
                                    "department_id": {
                                        "type": "integer"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jabatan"
                ],
                "summary": "Delete jabatan",
                "operationId": "afc61976bb8e1c758dc9ceec6b6388cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jabatans/{id}/status": {
            "patch": {
                "tags": [
                    "Jabatan"
                ],
                "summary": "Toggle jabatan status",
                "operationId": "5a9ac095db76b94a7ff1ad37bf31b70d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}